gvfs-helper-client: remove odb check
This check verifies that the ODB matches what we supplied, but there are
some subtleties around Windows path names that can cause inexact matches
to be logically the same. Since this check is really intended only for
debugging and development purposes, let's remove it for now as a quick
workaround.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>