Optionally include `scalar` when building/testing Git
The Scalar project started out as a separate project. As designated
successor of VFS for Git, its goal has been from the start to apply the
learnings from the VFS for Git project to allow Git to scale. In
contrast to VFS for Git, Scalar very much wants Git to be aware that the
checkout is sparse and the clone is partial (as opposed to VFS for Git
which tried to pretend that everything is there via a virtual file
system driver).

The intention has always been to integrate as much as possible of Scalar
into Git proper. And this has been the case: the partial clone feature,
the sparse checkout (cone mode), the commit graph, multi-pack index
files, scheduled maintenance tasks, all of these wonderful things can
now be enjoyed by Git users _without_ having to install Scalar.

So what remains of Scalar was ported to C, and put into
`contrib/scalar/`. And to make it easier for folks to build and install
it along with Git, this here patch makes it possible to simply drop
`INCLUDE_SCALAR = YesPlease` into the `config.mak` file to make it so.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>