fixup! scalar: add the `cache-server` command
Get rid of the cmdmodes (`--set` _really_ needs to take an argument, for
example, it is not a command mode).

Also, the `--list` functionality was not even implemented.

Further, the original `scalar.exe` accepts an optional `<worktree>`
argument (even if it calls it "enlistment", but that's not really Git
terminology, it sounds more like Perforce).

Finally, we need to look up the URL if we try to access the remote via a
nick name; Otherwise `supports_gvfs_protocol()` would think that the
URL's protocol isn't `https` and return early, without doing anything.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>