fixup! Implement `scalar diagnose`
Let's really include the full Git version in the `diagnose` output.

It is unclear why there is a link error when calling
`get_version_info()` and `load_builtin_commands()` is missing (nothing
in `get_version_info()` comes even close to building the command list),
but providing a dummy implementation of the latter function works around
it.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>