fixup! scalar register: set recommended config settings
We need to enforce the distinction between an enlistment and a worktree.
A Scalar enlistment is the top-level directory in which the `src/`
directory lives, which is a Git worktree.

The idea is to encourage as few untracked files within a worktree as
possible.

The `scalar register` command expects an optional enlistment path, not
a worktree path.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>