scalar clone: fix `dir_inside_of()` on case-insensitive filesystems
The `dir_inside_of()` function relies on `ignore_case` being set
correctly. This global variable is set only when reading the config, and
only when it encounters the `core.ignoreCase` key. This key (and for
that matter, the repository config) is only initialized once a
repository exists...

Therefore, let's first initialize the repository, and read the config,
before calling `dir_inside_of()`.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>