fixup! scalar register: set recommended config settings
Let's drop the `core.filemode` setting: `git init` already initializes
this correctly.

And then, disable the untracked cache on Windows. This hurts a bit
because the performance improvement is notable, but we have to turn it
off because it is apparently unreliable: the functional tests
consistently fail with `core.untrackedCache = true`.

Signed-off-by: Johannes Schindelin <johasc@microsoft.com>