fixup! scalar: add the `cache-server` command
Since we're planning on introducing the `delete`, `help` and `version`
commands earlier than the `cache-server` command (the former are in the
upstreamable part of `contrib/scalar/`, the latter is not), it makes
sense to have that fact reflected in the order of the implementation,
the documentation and the testing.

This commit is best viewed using `--color-moved`.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>