fixup! scalar clone: support GVFS-enabled remote repositories
Before figuring out whether we can use the GVFS protocol, make sure that
we ask the credential helper to use the URL path as part of the key when
accessing Azure Repos.

Also, the GVFS protocol only works via HTTP/1.1, so let's force that
once we know that we want to use that protocol.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>