fixup! gvfs-helper: add prefetch support
Due to how we are calling 'git index-pack' and piping stdout, I found an
instance where our 'git index-pack' subcommand is timing out. Since we
are using pipe_command() to get the stdout and use that hash to create
a new filename, the 'ip.stderr = -1' line breaks the expected behavior,
especially when providing '-v'.

Let's be more careful about stderr when cloning.

The case I saw was for using the prefetch endpoint on Azure Repos on a
repo that doesn't have configured cache servers. The pack-file was over
1GB so the progress indicator reliably hit a problem.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>