scalar: only retry a full clone if not using GVFS protocol
I found this error in a build, and it was confusing: the call to
set_config() to un-set the partial clone settings failed because those
settings didn't exist. We were in the GVFS protocol case instead. This
will at least give us a clearer reason why such a fetch fails.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>