sparse-index: recompute cache-tree
This took WAY TOO LONG to figure out, but when we disable the sparse
index for a given command we can get in a state where the in-memory
cache tree is actually equal to the sparse index's cache tree, resulting
in incorrect entry_count values. By clearing the cache tree before
converting to sparse, we avoid this issue.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>