sparse-index: create expand_to_pattern_list()
In order to allow modifying the sparse-checkout cone using a sparse
index without expanding to a full one, we need to be able to replace
sparse directory entries with their contained files and subdirectories
so other code paths can discover those cache entries and write the
corresponding files to disk before committing the index.

We already have logic in ensure_full_index() that expands the index
entries, so we will use that as our base. Create
expand_to_pattern_list() which takes a pattern list, but for now mostly
ignores it. The current implementation is only correct when the pattern
list is NULL as that does the same as ensure_full_index(). In fact,
ensure_full_index() is converted to a shim over
expand_to_pattern_list().

A future update will actually implement expand_to_pattern_list() to its
full capabilities. For now, it is created and documented. We also start
using doc-style comments in sparse-index.h.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>