checkout-index: expand sparse checkout compatibility tests
Add tests to cover `checkout-index` cases related to sparse checkouts: files
and folder, both inside and outside of sparse checkout definition. New tests
will serve as a baseline for expected behavior and performance when
integrating `checkout-index` with the sparse index.

Of note is the test demonstrating the behavior of `checkout-index --all`;
creating files even outside the sparse checkout definition is somewhat
unintuitive, and will cause significant performance issues when run with a
sparse index. The goal of later changes will be to change this default
behavior and introduce a `--sparse` flag to enable populating files outside
the sparse definition.

Signed-off-by: Victoria Dye <vdye@github.com>