update-index: block directories in `--cacheinfo`
Add explicit check for directory entry mode, throwing an error if found.
Non-sparse indexes are assumed and expected to contain no sparse directory
entries; if introduced, index corruption and BUGs appear when other commands
utilize the index.

Note that, although sparse directories could theoretically be added with
`--cacheinfo` in a sparse index, `git update-index` is not yet integrated
with the sparse index. Implementation of sparse directory compatibility is
therefore deferred until that integration is complete.

Signed-off-by: Victoria Dye <vdye@github.com>