checkout-index: integrate with sparse index
Add repository settings to allow usage of the sparse index. In order to prevent
unexpected errors when attempting to check out a sparse directory entry,
`checkout_file` directly checks whether a found entry is a sparse directory
and, if so, exits with an error. The test corresponding to this case now
verifies the error message, intentionally differing from the non-sparse index
scenarios.

Signed-off-by: Victoria Dye <vdye@github.com>