merge-recursive: add merge function arg to `merge_recursive_generic`
The `merge_recursive_generic` function is a wrapper that allows running a merge
on trees (rather than commits) by wrapping the trees as virtual commits. This
functionality is valuable not only for use with `merge_recursive`, but also
`merge_ort_recursive`. Having the merge function used by
`merge_recursive_generic` be an argument allows greater flexibility of usage.

Signed-off-by: Victoria Dye <vdye@github.com>