read-tree: use `merge-ort` for `twoway_merge` for mismatched sparse dirs
If an edit/edit conflict exists in a sparse directory, it may be a "fake"
conflict (in that the edits may have been to different files, in which case
there is no conflict). When this case occurs in a `read-tree` invocation with
two input trees, the sparse directories are merged with `merge-ort` - if any
conflicts are found, they are reported as a `reject_merge`.

Signed-off-by: Victoria Dye <vdye@github.com>