reset: make non-`--mixed` options sparse-aware
In order to accurately reconstruct the cache tree in `prime_cache_tree_rec`,
the function must determine whether the currently-processing directory in the
tree is sparse or not. If it is not sparse, the tree is parsed and subtree
recursively constructed. If it is sparse, no subtrees are added to the tree and
the entry count is set to 1 (representing the sparse directory itself).

Signed-off-by: Victoria Dye <vdye@github.com>