update-microsoft-git: interface with apt-get on Linux
The 'git update-microsoft-git' builtin already interfaces with the
git-for-windows/git upgrader on Windows and brew on macOS. Add an
integration with apt-get on Ubuntu. The implementation attempts to call
apt-get on all Linux machines, but it will report a failure when that is
not available. The two commands are:

1. sudo apt-get update
2. sudo apt-get upgrade microsoft-git

The first has an error message because that will likely only fail when
apt-get does not exist on the path (or the user fails to supply the sudo
password). The error condition for the second is passed to the user. It
is possible that a user installs a package manually and not through a
Microsoft feed, in which case the 'upgrade' command will fail with an
error message.

Extra newlines are added to provide whitespace between the Git messages
and the dense output from apt-get.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>