fetch/pull: use the sparse index
The 'git fetch' and 'git pull' commands somehow read the index. In the
case of 'git pull', much of the reason is that the HEAD might change due
to the update from the remote. It is unclear why 'git fetch' needs it.
However, I was able to measure time in ensure_full_index() because these
commands do parse an index.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>