sparse-checkout: block subcommands when disabled
Exit `git sparse-checkout` with error when executing subcommands `add`,
`set`, `reapply`, and `list` with sparse-checkout disabled. Update tests in
`t1091-sparse-checkout-builtin.sh` to avoid this error when not explicitly
tested, and add a new test to directly verify the error.

This is intended to eliminate cases in which sparse-checkout can be
initialized and/or reapplied outside of `git sparse-checkout init`. Blocking
execution of the subcommands is chosen over adding options to allow them to
perform full-featured `init`. Because `git sparse-checkout init` and `git
sparse-checkout disable` are established as explicit standalone operations,
it is potentially confusing to have "alternative" paths to initializing the
sparse checkout. This confusion was directly observed in off-team testing of
sparse-checkout when a user - not expecting re-initialization with `git
sparse-checkout add` - ended up with corrupted cone-mode patterns after
running `git sparse-checkout init --cone`.

Note that this reverses the decision to have `git sparse-checkout set`
implicitly enable the sparse-checkout in f6039a9423 (sparse-checkout: 'set'
subcommand, 2019-11-21). The `t1091` test 'set enables config' is removed
accordingly.

Signed-off-by: Victoria Dye <vdye@github.com>