status: fix nested sparse directory diff in sparse index
Add the `recursive` flag to `wt_status_collect_changes_index`. In order for
the error to occur, a number of conditions needed to be met:

* sparse index is enabled
* there is a difference between the index and HEAD in a file *more than* one
  level deep in a sparse directory
* the sparse directory index entry is *not* expanded

In this scenario, `git status` would not recurse into the sparse directory's
subdirectories to identify which file contained the difference between the
index and HEAD. Instead, it would report the immediate subdirectory itself
as "modified".

Example:

$ git init
$ mkdir -p sparse/sub
$ echo test >sparse/sub/foo
$ git add .
$ git commit -m "commit 1"
$ echo somethingelse >sparse/sub/foo
$ git add .
$ git commit -a -m "commit 2"
$ git sparse-checkout set --cone --sparse-index 'sparse'
$ git reset --soft HEAD~1
$ git status
On branch master
You are in a sparse checkout.

Changes to be committed:
  (use "git restore --staged <file>..." to unstage)
        modified:   sparse/sub

This is corrected by enabling the `recursive` diff option in
`wt_status_collect_changes_index`. Given the same repository setup as the
example above, the result of `git status` is now:

$ git status
On branch master
You are in a sparse checkout.

Changes to be committed:
  (use "git restore --staged <file>..." to unstage)
        modified:   sparse/sub/foo

Signed-off-by: Victoria Dye <vdye@github.com>