sparse-index: prevent repo root from becoming sparse
Prevent the repository root from being collapsed into a sparse directory by
considering an empty path "inside the sparse-checkout". When collapsing a
sparse index (e.g. in `git sparse-checkout reapply`), the root directory
typically could not become a sparse directory due to the presence of in-cone
child files/directories. However, if no such in-cone files/directories were
present, there was no explicit check signaling that the "repository root
path" (an empty string, in the case of 'convert_to_sparse(...)') was
in-cone, and a sparse directory index entry would be created from the
repository root directory.

The documentation in Documentation/git-sparse-checkout.txt explicitly states
that the files in the root directory are expected to be in-cone for a
cone-mode sparse-checkout. Collapsing the root into a sparse directory entry
violates that assumption, as sparse directory entries are implicitly
expected to be outside the sparse cone and have `skip-worktree` enabled,
manifesting in unexpected behavior in e.g. 'git status'.

Treating an empty (root) path as in-cone prevents the creation of a root
sparse directory in 'convert_to_sparse(...)'. Because the repository root is
otherwise never compared with sparse patterns (in both cone-mode and
non-cone sparse-checkouts), the new check does not cause additional changes
to how sparse patterns are applied.

Signed-off-by: Victoria Dye <vdye@github.com>